The changes are now extensive enough that I'm employing a separate branch --- reusing the old CPT directory. Distinct branches allow me to compare the operation of the original version with this version during debugging.
The behavior of this exe is no longer normal as it was in the previous 3 revisions, but enormously stripped-down.  Still, it is now able to open files and draw connects, areas, and parts to the screen.  But essentially all user input is disabled.  Large numbers of #ifdef's are used to hide older incompatible code from the compiler.

There are a lot of added files for this commit, so hopefully none are missing.

+++ branches/cpt/DlgIvex.cpp
@@ -97,7 +97,7 @@ void CivexDlg::OnBnClickedButtonInfile()
 	// get project file name
 	CString s ((LPCSTR) IDS_IvexLibraryFiles);
 	CFileDialog dlg( 1, "mod", LPCTSTR(""), 0, s );
-	int err = (int)dlg.DoModal();
+	int err = dlg.DoModal();
 	if( err == IDOK )
 	{
 		in_pathname = dlg.GetPathName();
